Move PDF related files in own folder + refactor
Move PDF related files in own folder
Refactor PDF file import functions
